+++ R/wrangling_issues.R
@@ -50,19 +50,13 @@ new_issue <- function(title,
         issue <- list(title = title,
                       body = body,
                       number = as.integer(number),
-                      created_at = created_at |>
-                          as.POSIXct(origin = "1970-01-01") |>
-                          as.integer() |>
-                          as.POSIXct(origin = "1970-01-01"),
+                      created_at = format_timestamp(created_at),
                       labels = labels,
                       milestone = milestone,
                       repo = repo,
                       owner = owner)
     } else if (!missing(issue)) {
-        issue[["created_at"]] <- issue[["created_at"]] |>
-            as.POSIXct(origin = "1970-01-01") |>
-            as.integer() |>
-            as.POSIXct(origin = "1970-01-01")
+        issue[["created_at"]] <- format_timestamp(issue[["created_at"]])
         issue[["number"]] <- as.integer(issue[["number"]])
     }
     class(issue) <- "IssueTB"